ASCII printing of floating point numbers must cast to (double) to support quad precision

+++ src/vec/vec/examples/tests/ex36.c
@@ -60,7 +60,7 @@ int main(int argc,char **argv)
   for (i = 0; i<m; i++) {
     ierr = PetscSynchronizedPrintf(PETSC_COMM_WORLD,
                                    "%d: idx[%D] == %D; val[%D] == %f\n",
-                                   rank, i, indices[i], i, PetscRealPart(values[i]));CHKERRQ(ierr);
+                                   rank, i, indices[i], i, (double)PetscRealPart(values[i]));CHKERRQ(ierr);
   }
   ierr = PetscSynchronizedFlush(PETSC_COMM_WORLD,PETSC_STDOUT);CHKERRQ(ierr);
 
@@ -94,7 +94,7 @@ int main(int argc,char **argv)
   ierr = PetscSynchronizedPrintf(PETSC_COMM_WORLD, "%d: Fetched values:\n", rank);CHKERRQ(ierr);
   for (i = 0; i<m; i++) {
     ierr = PetscSynchronizedPrintf(PETSC_COMM_WORLD, "%d: idx[%D] == %D; val[%D] == %f\n",
-                                   rank, i, indices[i], i, PetscRealPart(values[i]));CHKERRQ(ierr);
+                                   rank, i, indices[i], i, (double)PetscRealPart(values[i]));CHKERRQ(ierr);
   }
   ierr = PetscSynchronizedFlush(PETSC_COMM_WORLD,PETSC_STDOUT);CHKERRQ(ierr);
